pyproject.toml
[project]
name = "place-research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "beautifulsoup4>=4.13.4",
    "click>=8.2.1",
    "fastapi>=0.115.0",
    "googlemaps>=4.10.0",
    "jwt>=1.4.0",
    "osmnx>=2.0.6",
    "pandas>=2.3.2",
    "passlib[bcrypt]>=1.7.4",
    "pwdlib[argon2]>=0.3.0",
    "pydantic[email]>=2.11.7",
    "pydantic-settings>=2.7.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-json-logger>=3.2.1",
    "python-multipart>=0.0.21",
    "redis>=7.1.0",
    "requests>=2.32.5",
    "rich>=14.1.0",
    "sqlalchemy>=2.0.45",
    "sqlmodel>=0.0.31",
    "uvicorn[standard]>=0.34.0",
]

[project.scripts]
research = "place_research.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv]
package = true

[dependency-groups]
dev = [
    "bandit>=1.8.6",
    "interrogate>=1.7.0",
    "isort>=6.0.1",
    "pre-commit>=4.3.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.12.10",
]
jupyter = ["jupyter>=1.1.1"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
